Copy SqlRowMerger from internal code
This moves things mostly as is. I will move everything to appropriate packages in a follow on (e.g. utils will move to internal). Wanted to keep the changes as minimal as possible

Renames ProtoRowsBatchMergingStateMachine to ProtoRowsMergingStateMachine and removes the obsolete ProtoRowsMergingStateMachine

This simplifies the internal code by removing support for the old api. It should not change any behavior for he ProtoRowsBatch API

Change-Id: I8e8267a9cfeb59e0f8455876d1bccdb7c3778b76