Implement proto to java conversion via AbstractProtoStructReader
This will be re-used for structs in the following commit. The current
design implements this at the Row/Struct level instead of ResultSet &
Struct so that result set can operate on rows of different types when we
support arrow. ResultSet will then be a light wrapper where each getType
call is passed through to the current row.

Change-Id: I21fb4cc0cd830110d5455581491c5b1861efbcb1